requirements.txt
streamlit>=1.24.0
torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0